Simulation No Longer Resetting Properly
The simulation runs fine, but since creating a simulation state object,
there are problems with the simulation not resetting when you try to
run it again.

I'm considering writing the un-run simulation to a file and completely
reloading it as a solution… but I know jack shit about file handling.